requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
bcrypt==4.3.0
cffi==2.0.0
click==8.2.1
cryptography==46.0.1
Django==5.0.7
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
et-xmlfile==1.1.0
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
numpy==2.1.0
openpyxl==3.1.5
pandas==2.2.2
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
pytz==2024.1
rsa==4.9.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2024.1
uvicorn==0.35.0
